Rename command module + library
