+++ library/command/qsl_command/build.gradle
@@ -1,0 +1,12 @@
+plugins {
+	id("qsl.module")
+}
+
+qslModule {
+	moduleName = "qsl_command"
+	version = "1.0.0"
+	library = "commands"
+	coreDependencies([
+			"qsl_base"
+	])
+}
