+++ library/command/qsl_command/src/main/java/org/quiltmc/qsl/command/mixin/HelpCommandAccessor.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2021 QuiltMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.quiltmc.qsl.command.mixin;
+
+import com.mojang.brigadier.exceptions.SimpleCommandExceptionType;
+import net.minecraft.server.command.HelpCommand;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(HelpCommand.class)
+public interface HelpCommandAccessor {
+	@Accessor("FAILED_EXCEPTION")
+	static SimpleCommandExceptionType getFailedException() {
+		throw new AssertionError("mixin");
+	}
+}
