+++ library/command/qsl_command/src/main/java/org/quiltmc/qsl/command/mixin/client/ClientPlayerEntityMixin.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2021 QuiltMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.quiltmc.qsl.command.mixin.client;
+
+import net.minecraft.client.network.ClientPlayerEntity;
+import org.quiltmc.qsl.command.impl.client.ClientCommandInternals;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(ClientPlayerEntity.class)
+abstract class ClientPlayerEntityMixin {
+	@Inject(method = "sendChatMessage", at = @At("HEAD"), cancellable = true)
+	private void onSendChatMessage(String message, CallbackInfo info) {
+		if (ClientCommandInternals.executeCommand(message)) {
+			info.cancel();
+		}
+	}
+}
