+++ library/command/qsl_command/src/main/resources/quilt_command.mixins.json
@@ -1,0 +1,17 @@
+{
+  "required": true,
+  "package": "org.quiltmc.qsl.command.mixin",
+  "compatibilityLevel": "JAVA_16",
+  "mixins": [
+    "CommandManagerMixin",
+    "HelpCommandAccessor",
+    "client.ClientCommandSourceMixin",
+    "client.ClientPlayerEntityMixin",
+    "client.ClientPlayNetworkHandlerMixin",
+    "client.MinecraftClientMixin"
+  ],
+  "client": [],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
