--- library/commands/build.gradle
@@ -1,8 +1,0 @@
-plugins {
-	id("qsl.library")
-}
-
-qslLibrary {
-	libraryName = "commands"
-	version = "1.0.0"
-}
