--- library/commands/qsl_commands/build.gradle
@@ -1,12 +1,0 @@
-plugins {
-	id("qsl.module")
-}
-
-qslModule {
-	moduleName = "qsl_commands"
-	version = "1.0.0"
-	library = "commands"
-	coreDependencies([
-			"qsl_base"
-	])
-}
