--- library/commands/qsl_commands/src/main/java/org/quiltmc/qsl/commands/api/client/ClientCommandRegistrationCallback.java
@@ -1,47 +1,0 @@
-/*
- * Copyright 2021 QuiltMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.quiltmc.qsl.commands.api.client;
-
-import com.mojang.brigadier.CommandDispatcher;
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
-import org.quiltmc.qsl.base.api.event.ArrayEvent;
-
-/**
- * Event for registering client-side commands.
- *
- * @see ClientCommandManager
- */
-@FunctionalInterface
-@Environment(EnvType.CLIENT)
-public interface ClientCommandRegistrationCallback {
-	/**
-	 * Invoked when client-side commands are registered.
-	 */
-	ArrayEvent<ClientCommandRegistrationCallback> EVENT = ArrayEvent.create(ClientCommandRegistrationCallback.class, callbacks -> dispatcher -> {
-		for (var callback : callbacks) {
-			callback.registerCommands(dispatcher);
-		}
-	});
-
-	/**
-	 * Called when client-side commands are registered.
-	 *
-	 * @param dispatcher the command dispatcher.
-	 */
-	void registerCommands(CommandDispatcher<QuiltClientCommandSource> dispatcher);
-}
