--- library/commands/qsl_commands/src/main/java/org/quiltmc/qsl/commands/mixin/client/MinecraftClientMixin.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2021 QuiltMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.quiltmc.qsl.commands.mixin.client;
-
-import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.RunArgs;
-import org.quiltmc.qsl.commands.impl.client.ClientCommandInternals;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-
-@Mixin(MinecraftClient.class)
-abstract class MinecraftClientMixin {
-	@Inject(method = "<init>", at = @At("RETURN"))
-	private void onConstruct(RunArgs args, CallbackInfo info) {
-		ClientCommandInternals.initialize();
-	}
-}
