--- library/commands/qsl_commands/src/main/resources/fabric.mod.json
@@ -1,24 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "quilt_commands",
-  "name": "Quilt Command API",
-  "version": "${version}",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "icon": "assets/quilt_base/icon.png",
-  "contact": {
-    "homepage": "https://quiltmc.org",
-    "issues": "https://github.com/QuiltMC/quilt-standard-libraries/issues",
-    "sources": "https://github.com/QuiltMC/quilt-standard-libraries"
-  },
-  "authors": [
-    "QuiltMC"
-  ],
-  "depends": {
-    "quilt_loader": ">=0.13.1-rc.10"
-  },
-  "description": "Quilt APIs relating to commands.",
-  "mixins": [
-    "quilt_commands.mixins.json"
-  ]
-}
