--- library/commands/qsl_commands/src/main/resources/quilt_commands.mixins.json
@@ -1,17 +1,0 @@
-{
-  "required": true,
-  "package": "org.quiltmc.qsl.commands.mixin",
-  "compatibilityLevel": "JAVA_16",
-  "mixins": [
-    "CommandManagerMixin",
-    "HelpCommandAccessor",
-    "client.ClientCommandSourceMixin",
-    "client.ClientPlayerEntityMixin",
-    "client.ClientPlayNetworkHandlerMixin",
-    "client.MinecraftClientMixin"
-  ],
-  "client": [],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
