--- library/commands/qsl_commands/src/testmod/resources/fabric.mod.json
@@ -1,27 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "quilt_commands_api_testmod",
-  "name": "Quilt Command API Test Mod",
-  "version": "${version}",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "icon": "assets/quilt_base/icon.png",
-  "contact": {
-    "homepage": "https://quiltmc.org",
-    "issues": "https://github.com/QuiltMC/quilt-standard-libraries/issues",
-    "sources": "https://github.com/QuiltMC/quilt-standard-libraries"
-  },
-  "authors": [
-    "QuiltMC"
-  ],
-  "depends": {
-    "quilt_loader": ">=0.13.1-rc.10"
-  },
-  "entrypoints": {
-    "main": [
-      "org.quiltmc.qsl.commands.test.CommandsApiTest"
-    ]
-  },
-  "description": "Testmod for Quilt APIs relating to commands.",
-  "mixins": []
-}
